refactor : change name entity to entities
