+++ src/domain/entities/activity-entity.ts
