+++ src/domain/users/UsersRepository.ts
@@ -1,4 +1,4 @@
-import { User } from "../entity/user-entity";
+import { User } from "../entities/user-entity";
 
 export interface UserRepository {
   create(data: User): Promise<User>;
